big refactor of the UML generation, field support for structs roughed in
