--- queries/enums.scm
@@ -1,1 +1,0 @@
-(enum_item name: (type_identifier) @enum.name)
