+++ queries/newtype.rb
@@ -1,0 +1,4 @@
+class Query
+  class Newtype < Query
+  end
+end
