--- queries/structs.scm
@@ -1,1 +1,0 @@
-(struct_item name: (type_identifier) @struct.name)
