+++ queries/trait.scm
