+++ scripts/scm.rb
@@ -1,0 +1,6 @@
+class SCM < Query
+  def initialize(path)
+    super
+    @code = File.read(path)
+  end
+end
