+++ scripts/variant.rb
@@ -1,0 +1,3 @@
+
+class Variant
+end
